requirements.txt
amqp==5.2.0
anyjson==0.3.3
apturl==0.5.2
asgiref==3.7.2
async-timeout==4.0.3
bcrypt==3.2.0
billiard==4.2.0
blinker==1.4
Brlapi==0.8.3
celery==5.3.6
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.3
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-rest-swagger==2.2.0
django-simple-pagination==1.3
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-queryfields==1.1.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
duplicity==0.8.21
ephem==4.1.3
fasteners==0.14.1
filelock==3.6.0
future==0.18.2
gyp==0.1
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
inflection==0.5.1
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.3
keyring==23.5.0
kombu==5.3.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
openapi-codec==1.3.2
packaging==23.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
prompt-toolkit==3.0.43
protobuf==3.12.4
psycopg2==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-memcached==1.58
pytz==2024.1
pyxdg==0.27
PyYAML==6.0.1
redis==5.0.2
reportlab==3.6.8
requests==2.31.0
SecretStorage==3.3.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
systemd-python==234
typing_extensions==4.10.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.2.1
usb-creator==0.3.7
vine==5.1.0
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.13
xdg==5
xkit==0.0.0
zipp==1.0.0
